requirements.txt
# Usual dependencies:
Django==4.2
django-filter==23.1
django-storages==1.13.2
djangorestframework==3.14
requests==2.28.2
python-dotenv==1.0.0
psycopg2-binary==2.9.6
lxml==4.9.2
gunicorn==20.1.0

# Tests Specific:
coverage==7.2.3
responses==0.23.1
pylint==2.17.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
black==23.3.0
